Adopt changes of links and functions
