+++ tests/chainer_tests/functions_tests/rnn_tests/test_lstm.py
@@ -5,7 +5,7 @@ import numpy
 import chainer
 from chainer.backends import cuda
 from chainer import functions
-from chainer.functions.activation import slstm
+from chainer.functions.rnn import lstm
 from chainer import gradient_check
 from chainer import testing
 from chainer.testing import backend
@@ -23,15 +23,19 @@ def inject_backend_tests(method_names):
         testing.product({
             'use_cuda': [False],
             'use_ideep': ['never', 'always'],
-        }) +
+        })
         # GPU tests
-        [{'use_cuda': True}])
+        + [{'use_cuda': True}])
     return decorator
 
 
-@testing.parameterize(*testing.product({
+@testing.parameterize(*(testing.product({
+    'batch': [3, 2, 0],
+    'dtype': [numpy.float32],
+}) + testing.product({
+    'batch': [3],
     'dtype': [numpy.float16, numpy.float32, numpy.float64],
-}))
+})))
 @testing.fix_random()
 @inject_backend_tests([
     'test_forward',
@@ -43,25 +47,26 @@ def inject_backend_tests(method_names):
     'test_no_gh_backward',
     'test_flat_no_gh_backward',
     'test_double_backward'])
-class TestSLSTM(unittest.TestCase):
+class TestLSTM(unittest.TestCase):
 
     def setUp(self):
-        c_prev1 = numpy.random.uniform(-1, 1, (3, 2, 4)).astype(self.dtype)
-        c_prev2 = numpy.random.uniform(-1, 1, (3, 2, 4)).astype(self.dtype)
-        x1 = numpy.random.uniform(-1, 1, (3, 8, 4)).astype(self.dtype)
-        x2 = numpy.random.uniform(-1, 1, (3, 8, 4)).astype(self.dtype)
+        dtype = self.dtype
+        hidden_shape = (3, 2, 4)
+        x_shape = (self.batch, 8, 4)
+        y_shape = (self.batch, 2, 4)
 
-        gc = numpy.random.uniform(-1, 1, (3, 2, 4)).astype(self.dtype)
-        gh = numpy.random.uniform(-1, 1, (3, 2, 4)).astype(self.dtype)
+        c_prev = numpy.random.uniform(-1, 1, hidden_shape).astype(dtype)
+        x = numpy.random.uniform(-1, 1, x_shape).astype(dtype)
 
-        ggc_prev1 = numpy.random.uniform(-1, 1, (3, 2, 4)).astype(self.dtype)
-        ggc_prev2 = numpy.random.uniform(-1, 1, (3, 2, 4)).astype(self.dtype)
-        ggx1 = numpy.random.uniform(-1, 1, (3, 8, 4)).astype(self.dtype)
-        ggx2 = numpy.random.uniform(-1, 1, (3, 8, 4)).astype(self.dtype)
+        gc = numpy.random.uniform(-1, 1, hidden_shape).astype(dtype)
+        gh = numpy.random.uniform(-1, 1, y_shape).astype(dtype)
 
-        self.inputs = [c_prev1, c_prev2, x1, x2]
+        ggc = numpy.random.uniform(-1, 1, hidden_shape).astype(dtype)
+        ggx = numpy.random.uniform(-1, 1, x_shape).astype(dtype)
+
+        self.inputs = [c_prev, x]
         self.grad_outputs = [gc, gh]
-        self.grad_grad_inputs = [ggc_prev1, ggc_prev2, ggx1, ggx2]
+        self.grad_grad_inputs = [ggc, ggx]
 
         self.check_forward_options = {}
         self.check_backward_options = {'dtype': numpy.float64}
@@ -77,37 +82,37 @@ class TestSLSTM(unittest.TestCase):
         return [None if a is None else a[:, :, 0] for a in arrays]
 
     def forward_cpu(self, inputs):
-        c_prev1, c_prev2, x1, x2 = inputs
-        a1_in = x1[:, [0, 4]]
-        i1_in = x1[:, [1, 5]]
-        f1_in = x1[:, [2, 6]]
-        o1_in = x1[:, [3, 7]]
-        a2_in = x2[:, [0, 4]]
-        i2_in = x2[:, [1, 5]]
-        f2_in = x2[:, [2, 6]]
-        o2_in = x2[:, [3, 7]]
-        c_expect = _sigmoid(i1_in) * numpy.tanh(a1_in) + \
-            _sigmoid(i2_in) * numpy.tanh(a2_in) + \
-            _sigmoid(f1_in) * c_prev1 + \
-            _sigmoid(f2_in) * c_prev2
-        h_expect = _sigmoid(o1_in + o2_in) * numpy.tanh(c_expect)
+        c_prev, x = inputs
+        batch = x.shape[0]
+        a_in = x[:, [0, 4]]
+        i_in = x[:, [1, 5]]
+        f_in = x[:, [2, 6]]
+        o_in = x[:, [3, 7]]
+        c_expect = (_sigmoid(i_in) * numpy.tanh(a_in)
+                    + _sigmoid(f_in) * c_prev[:batch])
+        h_expect = _sigmoid(o_in) * numpy.tanh(c_expect)
         return c_expect, h_expect
 
     def check_forward(self, inputs, backend_config):
-        c_prev1, c_prev2, x1, x2 = inputs
-        c_expect, h_expect = self.forward_cpu(inputs)
+        # Compute expected out
+        c_prev, x = inputs
+        batch = x.shape[0]
+        c_expect_2 = c_prev[batch:]
+        c_expect_1, h_expect = self.forward_cpu(inputs)
 
         if backend_config.use_cuda:
             inputs = cuda.to_gpu(inputs)
         inputs = [chainer.Variable(xx) for xx in inputs]
 
         with backend_config:
-            c, h = functions.slstm(*inputs)
+            c, h = functions.lstm(*inputs)
             assert c.data.dtype == self.dtype
             assert h.data.dtype == self.dtype
 
         testing.assert_allclose(
-            c_expect, c.data, **self.check_forward_options)
+            c_expect_1, c.data[:batch], **self.check_forward_options)
+        testing.assert_allclose(
+            c_expect_2, c.data[batch:], **self.check_forward_options)
         testing.assert_allclose(
             h_expect, h.data, **self.check_forward_options)
 
@@ -124,7 +129,7 @@ class TestSLSTM(unittest.TestCase):
 
         with backend_config:
             gradient_check.check_backward(
-                functions.slstm, inputs, grad_outputs,
+                functions.lstm, inputs, grad_outputs,
                 **self.check_backward_options)
 
     def test_full_backward(self, backend_config):
@@ -162,39 +167,45 @@ class TestSLSTM(unittest.TestCase):
 
         with backend_config:
             gradient_check.check_double_backward(
-                chainer.functions.slstm, inputs, grad_outputs,
-                grad_grad_inputs, **self.check_double_backward_options)
+                chainer.functions.lstm, inputs, grad_outputs, grad_grad_inputs,
+                **self.check_double_backward_options)
 
-    def test_double_backward(self,  backend_config):
+    def test_double_backward(self, backend_config):
         self.check_double_backward(
             self.inputs, self.grad_outputs, self.grad_grad_inputs,
             backend_config)
 
 
-@testing.parameterize(*testing.product({
+@testing.parameterize(*(testing.product({
+    'batch': [3, 2, 0],
+    'dtype': [numpy.float32],
+}) + testing.product({
+    'batch': [3],
     'dtype': [numpy.float16, numpy.float32, numpy.float64],
-}))
+})))
 @testing.fix_random()
 @inject_backend_tests(['test_backward'])
-class TestSLSTMGrad(unittest.TestCase):
+class TestLSTMGrad(unittest.TestCase):
 
     def setUp(self):
-        c_prev1 = numpy.random.uniform(-1, 1, (3, 2, 4)).astype(self.dtype)
-        c_prev2 = numpy.random.uniform(-1, 1, (3, 2, 4)).astype(self.dtype)
-        x1 = numpy.random.uniform(-1, 1, (3, 8, 4)).astype(self.dtype)
-        x2 = numpy.random.uniform(-1, 1, (3, 8, 4)).astype(self.dtype)
-        c_next = numpy.random.uniform(-1, 1, (3, 2, 4)).astype(self.dtype)
+        hidden_shape = (3, 2, 4)
+        dtype = self.dtype
+        x_shape = (self.batch, 8, 4)
+        y_shape = (self.batch, 2, 4)
 
-        gc = numpy.random.uniform(-1, 1, (3, 2, 4)).astype(self.dtype)
-        gh = numpy.random.uniform(-1, 1, (3, 2, 4)).astype(self.dtype)
+        c_prev = numpy.random.uniform(
+            -1, 1, hidden_shape).astype(dtype)
+        x = numpy.random.uniform(-1, 1, x_shape).astype(dtype)
+        c_next = numpy.random.uniform(-1, 1, hidden_shape).astype(dtype)
 
-        ggc_prev1 = numpy.random.uniform(-1, 1, (3, 2, 4)).astype(self.dtype)
-        ggc_prev2 = numpy.random.uniform(-1, 1, (3, 2, 4)).astype(self.dtype)
-        ggx1 = numpy.random.uniform(-1, 1, (3, 8, 4)).astype(self.dtype)
-        ggx2 = numpy.random.uniform(-1, 1, (3, 8, 4)).astype(self.dtype)
+        gc = numpy.random.uniform(-1, 1, hidden_shape).astype(dtype)
+        gh = numpy.random.uniform(-1, 1, y_shape).astype(dtype)
 
-        self.inputs = [c_prev1, c_prev2, x1, x2, c_next, gc, gh]
-        self.grad_outputs = [ggc_prev1, ggc_prev2, ggx1, ggx2]
+        ggc_prev = numpy.random.uniform(-1, 1, hidden_shape).astype(dtype)
+        ggx = numpy.random.uniform(-1, 1, x_shape).astype(dtype)
+
+        self.inputs = [c_prev, x, c_next, gc, gh]
+        self.grad_outputs = [ggc_prev, ggx]
 
         self.check_backward_options = {'dtype': numpy.float64}
         if self.dtype == numpy.float16:
@@ -208,7 +219,7 @@ class TestSLSTMGrad(unittest.TestCase):
 
         with backend_config:
             gradient_check.check_backward(
-                slstm.SLSTMGrad(), inputs, grad_outputs,
+                lstm.LSTMGrad(), inputs, grad_outputs,
                 **self.check_backward_options)
 
     def test_backward(self, backend_config):
