+++ tests/chainer_tests/functions_tests/rnn_tests/test_n_step_rnn.py
