+++ tests/chainer_tests/links_tests/rnn_tests/test_mgu.py
