+++ tests/chainer_tests/links_tests/rnn_tests/test_n_step_lstm.py
