+++ tests/chainer_tests/links_tests/rnn_tests/test_tree_lstm.py
